Modify [expr.prim.lambda] wording to apply cleanly to C++14.
The previous version of this section had two paragraphs that were modified.
The first was removed in its entirety since the notion of "potentially
referenced" was introduced in P0588R1 which was post C++14. That paper
created the notion of "if the effect of any enclosing typeid expressions
([expr.typeid]) were ignored" - this did not exist before. As such,
reflexpr also does not need to be taken into account by the C++14 rebase.