Address Hubert's Table 11 footnote comment.
Hubert's Comment:

(minor) In Table 11, the reference to NOTE 1 is odd. I do not believe that
the example in the ISO/IEC Directives, Part 2 that has the numerals in
subscript intends to demonstrate a reference to a table note. Indeed, a
single note in a table is not to be numbered. I suggest using a footnote.